+++ src/main/java/me/ajlane/geo/repast/colonisation/csr/CompletelySpatiallyRandomParams.java
@@ -44,4 +44,32 @@ public class CompletelySpatiallyRandomParams {
         + "]";
   }
 
+  @Override
+  public int hashCode() {
+    final int prime = 31;
+    int result = 1;
+    long temp;
+    temp = Double.doubleToLongBits(baseRate);
+    result = prime * result + (int) (temp ^ (temp >>> 32));
+    temp = Double.doubleToLongBits(spreadRate);
+    result = prime * result + (int) (temp ^ (temp >>> 32));
+    return result;
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    if (this == obj)
+      return true;
+    if (obj == null)
+      return false;
+    if (getClass() != obj.getClass())
+      return false;
+    CompletelySpatiallyRandomParams other = (CompletelySpatiallyRandomParams) obj;
+    if (Double.doubleToLongBits(baseRate) != Double.doubleToLongBits(other.baseRate))
+      return false;
+    if (Double.doubleToLongBits(spreadRate) != Double.doubleToLongBits(other.spreadRate))
+      return false;
+    return true;
+  }
+
 }